Break out subtypes into modules, including arbitrary's
