--- zebra-chain/src/notes/tests.rs
@@ -1,1 +1,0 @@
-mod arbitrary;
